Changed astexport to just return the AST as a dictionary
- Cleaned up some unused code and imports
- Version
- Bugfixing

Fixed test
