+++ pydetector/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.13.0"
+__version__ = "0.14.0"
